Adds Luke McCarthy as additional copyright owner
This implementation of the ``inotify`` emitter
is based on the work of Luke McCarthy's
``fsmonitor`` library for Python. The original
implementation can be found at fsmonitor_

The original implementation does not comply with
our API specification, but also does not depend
on the pyinotify_ library. fsmonitor_ uses
Python ``ctypes`` to dynamically load ``inotify``
functionality.

.. _fsmonitor: http://github.com/shaurz/fsmonitor
.. _pyinotify: http://github.com/seb-m/pyinotify

Signed-off-by: Gora Khargosh <gora.khargosh@gmail.com>