Swallows fired but unpairable moved_from events
Two events with the same cookie may be notified
by inotify, but sometimes the key look up on the
cookie in the dictionary used to pair moved events
fails. I have yet to figure out why this happens.
As a test, you can try to execute ``git gc`` in
a recursively monitored directory at the root
level after disabling the KeyError handling.

I have marked this as a TODO in the code so we
do not forget about this issue.

Signed-off-by: Gora Khargosh <gora.khargosh@gmail.com>