Drops dependency on pywin32
Windows constants and API are accessible using
the Python ``ctypes`` module. ``pyfilesystem``
already contained the interface to the API we
need. Importing that code into the library
eliminates the dependency on ``pywin32``
which cannot be installed using ``pip``.

Dropping the dependency entirely made sense
to make it easier for users to install the
library.

Signed-off-by: Gora Khargosh <gora.khargosh@gmail.com>