Watchdog (on Windows) would create bad FS changed events streams
Code was stripping duplicate events from its internal queue. Example:
       Create X, Delete X, Create X, Edit X
Would be filtered to
       Create X, Delete X, Edit X
because the second "Create X" was still in the queue... thus it was
dropped. But that isn't a valid event stream. Can't 'Edit' a file that
doesn't exist.

Fixed by changing to use a queue that only dropped duplicate consecutive
events.