make 'watchmedo tricks' accept a positional argument to watch other dir
This provides an alternate mechanism for specifying what directory
should be watched for tricks.

tricks.yaml is a very valuable, user-friendly feature. However, many users
will want to keep it somewhere outside the project root, to keep it clean.
In addition, log and shell-command accept directory arguments and
specifying it is more explicit.

If an additional argument is not provided, you get the default behavior
of watching the directory containing the yaml file.