read_directory_changes: On directory create check if it may have been a move
On Windows if a folder is moved from outside the watched folder to inside it,
we only get a Directory Create from ReadDirectoryChangesW, no events for the directory's
children. Err on the side of caution and call them CreatedFile and DirEvents since we
can't be absolutely sure they came as the result of a move.