tricks: change files positional arg to --config
watchmedo tricks currently takes any number of positional arguments,
interpreting them as tricks.yaml files to source from.

With this patch, tricks.yaml files are specified with --config.
You can specify it multiple times for multiple tricks.yaml files.

If no --config is specified, it defaults to just "tricks.yaml".

This paves the way to interpret positional arguments as paths to watch.

That, in turn, would separate the function of tricks.yaml to specify what
should be done when, from the function of specifying what paths those things
should be done with.