Update dirsnapshot.py
Fix issue where one file was deleted and another file was created and the created file was assigned the inode of the deleted file.  This resulted in watchdog firing a moved event even though the files were totally unrelated.  Inodes can and will be recycled, it was happening frequently in our use case, depending solely on inode number to determine it is the same file won't work.  This change checks modification time before determining that the file with the given inode was moved instead of deleted/created.