Proposed fix for issue 392. The idea is that when you add handlers that are 'None' they really aren't handlers, and so it makes sense to start only the emitters but not the observer.
So this works, except that in the self test routines, it uses 'None' event handlers for
some reasons that I don't understand.  I made the tests work, but had to add a self.started
Boolean to the BaseObserver class in order to get 'is_alive' to work in one test.
Not sure that's the right answer, but it passes all the tests.