Adapt tests to improve BSD support (#633)
* Adapt tests to improve BSD support.

There was a false Linux|Windows dichotomy in some tests, so instead of
skipping certain tests on certain platforms, this now skips tests when
*not* on the expected platform.
E.g. InotifyFullEmitter tests shouldn't run on non-Linux and the
Windows-specific tests shouldn't run on non-Windows.

* Revert change to `test_separate_consecutive_moves`.

That way CI is happy and changes are conceptually more isolated.