[mac] Add support for non-recursive watches in FSEventsEmitter (#779)
* Add support for non-recursive watches in FSEventsEmitter

The underlying `fseventsd` service always observes recursively. As such we're adding a filter in `FSEventEmitter` to support the non-recursive behaviour that is supported by other emitters.

* Add more test cases

Test a few more scenarios for non-recursive watches. Do note that at least on macOS we'll get a `DirModifiedEvent` for the created subdirectory.

* Update changelog

* Limit new test scenarios to macOS

We're fixing a defect on macOS after all, and addressing the failures on other platforms looks like opening a can of worms that is not related to the original issue.

* Add tests for move events