[mac] Fix relative path handling for non-recursive watch in FSEventsObserver (#799)
* Add test for issue #797

This works for me locally, let's see what the CI machine says.

* Use absolute watch path in recursive check

A watch path may be relative, while an event path is always absolute. Since the check for recursive events simply compared the watch path with the event path then that would fail in case the watch was set to a relative folder. This resolves the issue.

Fixes #797

* Add support for ~ in watch path

Same problem as relative paths.

* Update changelog.rst

Co-authored-by: Mickaël Schoentgen <contact@tiger-222.fr>