use a single test/lint/hint/doc matrix
This started as just expanding the mypy matrix to cover platforms at least, and Python versions.  It grew to just being a general tox workflow avoiding the steps repetition.  It could be argued that docs and flake8 ought to be run against all the combinations to make sure they can be worked on in any of the situations.  But, I left that out at least for now as I'm not sure if that's a step beyond where this is wanted to go.

I noticed that the docs workflow specifically builds the PR branch instead of the regular automatic merge commit of the PR branch and the target branch.  Is `ref: ${{ github.event.inputs.branch }}` important?

If the overall change here is not preferred, I can make another PR that is more focused on expanding just the mypy matrix.