docs: Simplify the quickstart example (#1047)
Personally, I found the quickstart example to be jumbled and a little
confusing. It was hard for me to tell where I would put my own code to
react to file system events (indeed the quickstart never showed how to
do this).

I also think the selection of the path using sys.argv and place
everything under "if name == main" is unnecessary for the simple
example.