Use `select.poll()` if available.
Details:

As stated in the `select()` man page:
```
WARNING: select() can monitor only file descriptors numbers that are less than FD_SETSIZE
         (1024)—an unreasonably low limit for many modern applications—and this limitation will not
         change.
```

This can lead to `ValueError: filedescriptor out of range in select()` when using watchdog.
Following the advice of the `select()` man page, we use `select.poll()` instead, if available.
The call to `select()` used as a fallback.